Add comments to devtools, add IsDev, add IsPreRelease, Cleanup testing.
